added loader and loader manager
