+++ extensions/catalog/federated-catalog-cache/src/main/java/org/eclipse/dataspaceconnector/catalog/cache/crawler/Crawler.java
@@ -1,0 +1,93 @@
+package org.eclipse.dataspaceconnector.catalog.cache.crawler;
+
+import net.jodah.failsafe.RetryPolicy;
+import org.eclipse.dataspaceconnector.catalog.cache.spi.ProtocolAdapter;
+import org.eclipse.dataspaceconnector.catalog.cache.spi.UpdateRequest;
+import org.eclipse.dataspaceconnector.catalog.cache.spi.UpdateResponse;
+import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
+
+import java.util.List;
+import java.util.concurrent.BlockingQueue;
+
+import static java.lang.String.format;
+import static net.jodah.failsafe.Failsafe.with;
+
+public class Crawler implements Runnable {
+
+    private final List<ProtocolAdapter> adapters;
+    private final Monitor monitor;
+    private final BlockingQueue<UpdateResponse> queue;
+    private final RetryPolicy<Object> retryPolicy;
+
+    public Crawler(List<ProtocolAdapter> adapters, Monitor monitor, BlockingQueue<UpdateResponse> queue, RetryPolicy<Object> retryPolicy) {
+        this.adapters = adapters;
+        this.monitor = monitor;
+        this.queue = queue;
+        this.retryPolicy = retryPolicy;
+    }
+
+    @Override
+    public void run() {
+        adapters.forEach(adapter -> {
+            adapter.sendRequest(new UpdateRequest())
+                    .whenComplete((updateResponse, throwable) -> {
+                        if (throwable != null) {
+                            handleError(throwable);
+                        } else {
+                            handleResponse(updateResponse);
+                        }
+                    });
+        });
+    }
+
+    private void handleError(Throwable throwable) {
+        monitor.severe("An update-request failed", throwable);
+    }
+
+    private void handleResponse(UpdateResponse updateResponse) {
+        monitor.info(format("update-response received: %s", updateResponse.toString()));
+        var offered = with(retryPolicy).get(() -> queue.offer(updateResponse));
+        if (!offered) {
+            monitor.severe("Inserting update-response into queue failed due to timeout!");
+        }
+    }
+
+
+    public static final class Builder {
+        private List<ProtocolAdapter> adapters;
+        private Monitor monitor;
+        private BlockingQueue<UpdateResponse> queue;
+        private RetryPolicy<Object> retryPolicy;
+
+        private Builder() {
+        }
+
+        public static Builder newInstance() {
+            return new Builder();
+        }
+
+        public Builder adapters(List<ProtocolAdapter> adapters) {
+            this.adapters = adapters;
+            return this;
+        }
+
+        public Builder monitor(Monitor monitor) {
+            this.monitor = monitor;
+            return this;
+        }
+
+        public Builder queue(BlockingQueue<UpdateResponse> queue) {
+            this.queue = queue;
+            return this;
+        }
+
+        public Builder retryPolicy(RetryPolicy<Object> retryPolicy) {
+            this.retryPolicy = retryPolicy;
+            return this;
+        }
+
+        public Crawler build() {
+            return new Crawler(adapters, monitor, queue, retryPolicy);
+        }
+    }
+}
