+++ extensions/catalog/federated-catalog-cache/src/main/java/org/eclipse/dataspaceconnector/catalog/cache/spi/Loader.java
@@ -1,0 +1,7 @@
+package org.eclipse.dataspaceconnector.catalog.cache.spi;
+
+import java.util.Collection;
+
+public interface Loader {
+    void load(Collection<UpdateResponse> batch);
+}
