+++ extensions/catalog/federated-catalog-cache/src/main/java/org/eclipse/dataspaceconnector/catalog/cache/spi/LoaderManager.java
@@ -1,0 +1,7 @@
+package org.eclipse.dataspaceconnector.catalog.cache.spi;
+
+public interface LoaderManager {
+    void start();
+
+    void stop();
+}
