+++ extensions/catalog/federated-catalog-cache/src/main/java/org/eclipse/dataspaceconnector/catalog/cache/spi/UpdateResponse.java
@@ -1,0 +1,4 @@
+package org.eclipse.dataspaceconnector.catalog.cache.spi;
+
+public class UpdateResponse {
+}
