--- extensions/catalog/federated-catalog-cache/src/main/java/org/eclipse/dataspaceconnector/crawler/spi/UpdateRequest.java
@@ -1,4 +1,0 @@
-package org.eclipse.dataspaceconnector.crawler.spi;
-
-public class UpdateRequest {
-}
